Backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy==1.26.0
rasterio==1.3.8
tensorflow==2.15.0
Pillow==10.0.0
opencv-python==4.8.0.74
requests==2.31.0
geojson==3.2.0
setuptools==68.0.0 